requirements.txt
numpy-1.13.0
opencv-python-4.1.1
freetype-2.1.0